refactor: deepen command grammar
