+++ src/__tests__/semantic-grammar.test.ts
@@ -2,14 +2,12 @@ import { test } from 'vitest';
 import assert from 'node:assert/strict';
 import { DAEMON_COMMAND_GROUPS, PUBLIC_COMMANDS } from '../command-catalog.ts';
 import {
-  fillCommandCodec,
-  findCommandCodec,
-  interactionTargetCodec,
-  isCommandCodec,
-  longPressCommandCodec,
-  settingsCommandCodec,
-  waitCommandCodec,
-} from '../command-codecs.ts';
+  readFillTargetFromPositionals,
+  readInteractionTargetFromPositionals,
+  readLongPressTargetFromPositionals,
+  readSemanticInputFromCli,
+  readWaitOptionsFromPositionals,
+} from '../commands/semantic-grammar.ts';
 import type { CliFlags } from '../utils/command-schema.ts';
 
 const BASE_FLAGS: CliFlags = {
@@ -24,45 +22,45 @@ test('command catalog owns daemon routing groups', () => {
   assert.equal(DAEMON_COMMAND_GROUPS.replay.has(PUBLIC_COMMANDS.test), true);
 });
 
-test('wait codec preserves CLI bare text forms', () => {
-  const options = waitCommandCodec.decode(['Continue', '1500'], BASE_FLAGS);
+test('wait grammar preserves CLI bare text forms', () => {
+  const options = readWaitOptionsFromPositionals(['Continue', '1500'], BASE_FLAGS);
   assert.equal(options.text, 'Continue');
   assert.equal(options.timeoutMs, 1500);
 });
 
-test('interaction and fill codecs share ref, selector, and point grammar', () => {
-  assert.deepEqual(interactionTargetCodec.decode(['@e3', 'Email']), {
+test('interaction and fill grammar share ref, selector, and point parsing', () => {
+  assert.deepEqual(readInteractionTargetFromPositionals(['@e3', 'Email']), {
     ref: '@e3',
     label: 'Email',
   });
-  assert.deepEqual(fillCommandCodec.decode(['id=email', 'qa@example.com']), {
+  assert.deepEqual(readFillTargetFromPositionals(['id=email', 'qa@example.com']), {
     kind: 'selector',
     target: { selector: 'id=email' },
     text: 'qa@example.com',
   });
-  assert.deepEqual(fillCommandCodec.decode(['@e4', 'Email', 'qa@example.com']), {
+  assert.deepEqual(readFillTargetFromPositionals(['@e4', 'Email', 'qa@example.com']), {
     kind: 'ref',
     target: { ref: '@e4', label: 'Email' },
     text: 'qa@example.com',
   });
-  assert.deepEqual(fillCommandCodec.decode(['10', '20', 'hello']), {
+  assert.deepEqual(readFillTargetFromPositionals(['10', '20', 'hello']), {
     kind: 'point',
     target: { x: 10, y: 20 },
     text: 'hello',
   });
-  assert.deepEqual(longPressCommandCodec.decode(['@e4', '800']), {
+  assert.deepEqual(readLongPressTargetFromPositionals(['@e4', '800']), {
     ref: '@e4',
     durationMs: 800,
   });
-  assert.deepEqual(longPressCommandCodec.decode(['10', '20', '800']), {
+  assert.deepEqual(readLongPressTargetFromPositionals(['10', '20', '800']), {
     x: 10,
     y: 20,
     durationMs: 800,
   });
 });
 
-test('find and is codecs decode command action positionals', () => {
-  const findOptions = findCommandCodec.decode(['label', 'Continue', 'wait', '3000'], {
+test('find and is grammar decodes command action positionals', () => {
+  const findOptions = readSemanticInputFromCli('find', ['label', 'Continue', 'wait', '3000'], {
     ...BASE_FLAGS,
     platform: 'ios',
     findFirst: true,
@@ -74,17 +72,21 @@ test('find and is codecs decode command action positionals', () => {
   assert.equal(findOptions.timeoutMs, 3000);
   assert.equal(findOptions.first, true);
 
-  const isOptions = isCommandCodec.decode(['text', 'id=title', 'Welcome'], BASE_FLAGS);
+  const isOptions = readSemanticInputFromCli('is', ['text', 'id=title', 'Welcome'], BASE_FLAGS);
   assert.equal(isOptions.predicate, 'text');
   assert.equal(isOptions.selector, 'id=title');
   assert.equal(isOptions.value, 'Welcome');
 });
 
-test('settings codec owns positional grammar for CLI commands', () => {
-  const location = settingsCommandCodec.decode(['location', 'set', '37.3349', '-122.009'], {
-    ...BASE_FLAGS,
-    platform: 'ios',
-  });
+test('settings grammar owns positional parsing for CLI commands', () => {
+  const location = readSemanticInputFromCli(
+    'settings',
+    ['location', 'set', '37.3349', '-122.009'],
+    {
+      ...BASE_FLAGS,
+      platform: 'ios',
+    },
+  );
   assert.equal(location.platform, 'ios');
   assert.equal(location.setting, 'location');
   assert.equal(location.state, 'set');
